pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prs_connector_core"
version = "0.5.2"
description = "Базовый модуль для коннекторов платформы Пересвет"
authors = [
    {name = "ООО МПК", email = "info@mp-co.ru"},
]
readme = "README.md"
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "aiomqtt==2.3.2",
    "jsonata-python==0.5.3",
    "colorlog==6.9.0",
    "python-json-logger==3.3.0",
    "pydantic==2.11.3",
    "aiofiles==24.1.0",
    "ciso8601==2.3.2"
]

[project.urls]
Homepage = "https://github.com/mp-co-ru/prs-connector-core"
Repository = "https://github.com/mp-co-ru/prs-connector-core"
Documentation = "https://mp-co-ru.github.io/prs-connector-core/"

[tool.setuptools]
include-package-data = true
packages = ["prs_connector_core"]

[tool.setuptools.package-data]
"prs_connector_core" = ["py.typed"]
